requirements.txt
Django==1.11.18
django-jsonfield==1.0.1
pytz
